r972: Worked on the SyncTarget = Primary node case.
The resync LRU now also tracks pending read requests of the resycner
on the Target node. (As it is already done on the Source node).

It is also used to lock out application's updates to the storage
while the resync block is on the fly...

With this the recv_discard() became unused.