r1119: * Moved wake_up_interruptible(&mdev->ee_wait); into drbd_put_ee()
With drbd-0.7 this was the reason for a deadlock. On a
SyncSource node, it could happen that all EEs are used to
read data from the disk. In w_e_end_rsdata_req() there is just a
call to drbd_put_ee() but not call to wake_up_interruptible(&mdev->ee_wait);

And in the meantime the receiver thread went to sleep on ee_wait
because it did not got any EEs.

I think this is fixed now.