r1154: Filly fixed the "lc_get() failed!" Bug. What I did not see for all the time is:  We have a ussage pattern on an LRU cache where some elements have  a high reference count, but are otherwise idle.  A lot of other elements have a reference count of 0 but, and get'ed  and put'ed very frequently  -=>  The elements with the low reference count (actually 0 most of the time)  move to the front of the LRU list. The idle elements move to the back  of the LRU list.
Now, we have a third list: in_use
All elements with positive reference counts are on the in_use list.
All elements with a reference count of zero are on the LRU list, and may
be evicted.
Initially all elements are on the 'free' list. (And if something is
forced out with lc_del() it also goes to the 'free' list.)

No more "lc_get() failed!" messages!!! Yes!! Yes!