r1196: adding ap_bio_cnt
while the peers sometimes do not agree about the amount of data to be synced
and the receiver thread handing in drbd_disconnect because some counters seem to be counted wrong,
maybe adding yet an other counter helps to find the problem.

this adds atomic_t ap_bio_cnt, where ap stands for application.
it shows in /proc/drbd.

gets incremented *only* in drbd_make_request_common,
and gets decremented wherever and whenever we actually
give the bio_endio back to the upper layers.

no functionality changed this time.