r1659: The test for (rs_total != 0) was not okay. If sync was started with the "invalidate" command on the Secondary, it (rs_total) is set to a positive value, just after we send the BecomeSyncSource packet to the peer.
The peer(Primary) writes its whole bitmap and sets 
rs_total to a positive value afterwards. 

Writes that happen on the Primary while it writes its
bitmap are mirrored to the Secondary (where they are
mared as cean in the BitMap, since rs_total is already
set here), but when the ACK comes back to the Primary
rs_total is still 0 here.

With this patch that problem is fixed, they both consider
the sync to be started when the SYNC_STARTED bit is 
set on both node, and this happens with the first
ACK of the fir RSDATA packet... So the agree on the
same logic point in time.