r1661: Removed the "struct list_head *head" parameter from the drbd_process_ee() function, added a "be_sleepy" flag argument.
This this flag, the asender we avoid that the asender
thread eventually sleeps in drbd_process_ee().

This changes makes a strange lockup condition go
away... where receive and asender where sleeping
in drbd_process_ee(). 

I do not know exacle how the lockup begins, but
with this changes it can not lock-up.