r2013: There was a bug in the GC code that caused the GC's to diverge when there was one of the modifiers (--human, --timeout-expired, --do-what-I-say) used when making a node primary.
Usually this was not noticed, because there are lots of
events that caused the GCs to get in sync again. But at 
initialisation time (one has to use --do-what-I-say primary)
the bug could cause an attempt of a wrong sync (which was
actually prevented by some sanity code) if these commands
where issued on the primary: 
secondary, down, attach, primary, connect.

Fixed now.