r2465: Made a few improvements to the fault insertion test suite - still needs more work but it should test a wider range of things now. Right now, the trunk version does not pass these tests - it gets through inserting meta-data faults on read and write on the local node as 'drbdadm attach' is done but then fails pretty quickly if meta-data write faults on the remote node are inserted as attach is done - this results in the connection being torn down and I cant directly reconnect because of the issue with 'drbdadm connect' I reported earlier. If I do 'drbdadm disconnect' and then 'drbdadm connect', it works on the primary side (where I was doing the attach), but the other side gets into a Disconnecting state from which it doesn't recover (the connection was actually taken down when the fault occurred earlier on).
I'm still investigating the cause of this, but I suspect there is a hole
in the new Disconnecting state handling.

I also fixed a problem with the test agent code -- when it was spinning
waiting for the state to change to an expected value, it would run as
fast as possible reading the current state and comparing - I added a .5s
usleep.

Anyway - patch attached; will enhance again to automatically reconnect
as needed once we sort out the problem with connect failing.