r2601: Implemented a new after-slit-brain-0pri policy:        "discard-zero-changes"                       Auto sync from the node that modified                       blocks during the split brain situation, but only 		      if the target not did not touched a single block.                       If both nodes touched their data, this policy 		      falls back to disconnect.
And a new after-sb-1pri & 2pri policy
     "violently-as0p" Alsways take the decission of the "after-sb-0pri"
                      algorithm. Even if that causes case an erratic change
		      of the primarie's view of the data.
	              This is only usefull if you use an 1node FS (i.e.
		      not OCFS2 or GFS) with the allow-two-primaries
		      flag, _AND_ you really know what you are doing.
		      This is DANGEROUS and MAY CRASH YOUR MACHINE if you
		      have a FS mounted on the primary node.

To allow Simon and Ernest to run DRBD in this setup:

When running Primary/Primary if the Heartbeat connection goes down when
we recover we always split brain.  Simon had an idea which I have
implemented. He is on vacation  so this may not reflect his exact idea.

Essentially with this change, we do not create a new current UUID on the
node unless I/O is seen. This prevent Split-Brain mitigation when both
nodes are primary but only one node is originating I/O and never the
other.  He is only stand-by in that case.