r3006: Ernest pointed this out.
When we attach a disk to a connected device pair, we are in
disk state Nogotiating on the node that gets the disk attached.

When in the mean time some event on the peer node triggers
the transmittion of a state packet, we might end up using an
old p_uuid set and find a wrong resync decission.

Fixed that.