r3043: When you let DRBD connect to an listening TCP port and close that, then allowing it to connect a second time you trigger an workaround from the DRBD-0.7 days. It is printed to the syslog as "My msock connect got accepted onto peer's sock!". The the receiver sleeps for connect_int/2.
When you droped the network config during this time with
"drbdadm disconnect", you hit an OOPS.

The root of this bug was the the

if(drbd_request_state(mdev,NS(conn,WFConnection)) < SS_Success ) ...

statement in drbd_connect() elevated the connection state from
StandAlone to WFConnection. Later we dereference mdev->net_conf
and OOPS...

I fixed this by allowing that state change only when the connection
state before was >= Unconnected (I.e. we had a network config).