r3071: This fixes an issue Johannes triggered with his drbd-proxy.
The asender terminated while the connection state was WFReportParams,
actually it should set the connection state to NetworkFailure. -> 
That causes the receiver to exit really soon....

That if statement is a very old leftofer, making sure that
the fource NetworkFailure states to no overwrite each other. This
is nowadays done by the early state sanitizing code in 
_drbd_set_state() in a much better way.