make resync more robust, don't reset find bit offset too early.
by the sync group serialisation code,
a resync timer of a device may be rescheduled.

if it had already sent all requests (find reached the end of the bitmap),
but did not yet receive all the answers (some bits are still set),
it would re-request all those bits
if the start offset of the find gets reset too early.