keep requests in the transfer log until the next BarrierAck even for Protocol C
As suggested by Simon Graham:

On connection loss, any blocks corresponding to requests still in
the transfer log (i.e. not yet "barrier ack'ed") are dirtied in
the sync bitmap.
For protocol A and B, the necessity for this is obvious.
But even for protocol C we can benefit here, if the lower level
devices have volatile write-back cache, and the connection loss
was in fact a node crash of the Secondary.

Since a barrier request is expected to force pending requests to
stable storage, it is suficcient to dirty just the blocks
corresponding to open epochs.