State changes that happen while we are in WFReportParams have caused a lot of trouble.
We may not send state updates (ReportState) packets while we are
in WFReportParams state, because in that state the ReportState
packet has the special purpose to eventually start a resync
progress.

But during the WFReportParams state the current state gets
transmitted. We need to make sure that state changes that happen
between this transmission and the point in time we leave the
WFReportParams state get transmitted as well.

The changes necessary to fix this:

 1) Only use drbd_send_state() when state.conn >= Connected
    ( The only allowed and neccessary exception is the drbd_send_state()
      in drbd_connect() )
 2) Send a state update when leaving the WFReportParams state.