Schedule full sync in case we get connected to a diskless primary peer
The code to arrange a full sync in case a primary peer looses its
disk, has already been in place. Now also do the right thing in
case we realize fact while connecting.

Also made sure to not generate new current UUIDs over and over
in case we loos connection and reconnect in a loop.