Fixed a reace between the exit pathes the worker and receiver threads
In case we enter the conn == StandAlone && disk == Diskless state
there was a race condition between the exit pathes of the receiver
and the worker. In case the receiver exits its main loop first,
the receiver will wait forever for the WORK_PENDING bit to
get cleared. The worker will hang forever in drbd_mdev_cleanup.

Now the *only* place where wo go into StandAlone state is
in the exit path of the receiver. => This race can not happen
any longer.