Fixed a race in the resync finished vs sync-pause code path.
What's happening is that we finish syncing.
The secondary side goes to Connected at the end of
drbd_resync_finished()

The Primary side is still in drbd_resync_finished().  While still there,

the syncer isp flag gets set and unset due to a higher priority device
syncer getting paused/resumed.

Now the Primary side sends "Inconsistent" state to the Secondary since
we are still in drbd_resync_finished()
and have not gone Connected and UptoDate yet. But the other side is
connected and UpToDate though.

The secondary goes WFBitMapS via drbd_sync_handshake()

In the mean time the Primary finally goes Connected and UptoDate.

The Secondary is stuck in WFBitMapS.