Close window where old state can be sent by drbd_send_state()
A previous change made this routine acquire the drbd_state_lock() to
ensure that it wont be sent when there is a cluster wide state change
in progress. Unfortunately, the code in drbd_request_state released
the lock before updating the local state which means that it was still
possible for drbd_send_state() to send stale data.

This fix sets the local state whilst still holding the state lock.