Eliminated drbd_thread_signal().
drbd_thread_signal() could cause an OOPS when it got into an race
with drbd_thread_start(). (Thread state was != none, while task
was still NULL).
Decided to remove that function complete instead of fixing it.

It was possible to trigger an OOPS in drbd_thread_signal()
with:
 while :; do (set -e ; drbdadm down all ; drbdadm up all ); [[ $? = 0 ]] || break; done
when the peer node was down.