Added drbdadm support for -c config files and drbd event callbacks "handlers"
drbdadm now keeps track of which device minor was configured
in which config file. It maintains a symlink farm of
device minor -> config file mappings in /var/lib/drbd.

Now, when a kernel event (such as before-resync-target) occurs,
the list is scanned for the device minor and the appropriate config
file is used to look up the settings.

Up to now, if a handler was not configured in the "main" config file,
it was assumed to be unconfigured, giving "unexpected" results.