close race between drbd_thread_start and a thread in 'Exiting' t_state
Replaces the "hot fix" of commit f3404e227425392a61ca2d3a84df139f046eab4d
to close the race between thread Exiting and an other start attempt.

Testing the t_state (Exiting or Restarting) and assigning the new t_state
(None or Running) needs to be within the same spinlock (in drbd_thread_setup).
To be able to do that with minimal changes within the current drbd_thread
framework, I had to get rid of the task_mutex.
Changing of cpu_mask is now done in the respective "main loop" of all threads.