drbd thread code: move the stop-completion within the spinlock to avoid a race
exiting thread in drbd_thread_setup,  | thread start
 spin lock                            |
   t_state = None                     | spin lock
 spin unlock                          |   if (t_state == None)
                                      |    start new thread
                                      | spin unlock
                                      | wait_for_completion
 complete                             | oops. task not yet set,

as start and stop share one completion here,
and the exiting drbd_thread_setup
completion is taken to be the completion expected from
the new, starting drbd_thread_setup.
solution for now: move the completion within the spinlock,
but we have to get rid of this homegrown stuff.