big cleanup: whitespace, line breaks, braces, typedefs, comments, dead code; bring 8.0 more in line with 8.1/8.2, so cherry-picks become easier.
this is a reviewed "git diff drbd-8.0 drbd-8.1 | patch -p1",

to get the conflicts down when merging 8.0 -> 8.1,
and for cherry picks from 8.2 to 8.0,

this adjust the whitespace to match 8.1/8.2,
removes dead code,
removes the typedefs,
removes volatile,
removes DEVICE_NAME
uses DRBD_MAJOR

this makes 8.1 obsolete - which it is, meanwhile, anyways.

sorry for all the noise.