Allow DRBD to user other cn_idx values in case our default value is already taken.
It turned out that the CN_IDX_DRBD value we used got a user in the main
line kernel in the mean time. Now one can set the cn_idx value as
module parameter, also the module tries to find a free value on its own,
and the userland (drbdsetup) gets the value via sysfs.