Sane resync after detach on primary
Up to now we forced a full resync after a detach on a primary node
happened (no matter if it is caused by an administrative command,
or by an IO error).

Nowadays there is actually no reason to treat this worse than a crash
of a primary node.

When a primary crashes, we apply the AL afterwards to the bitmap.
The same is sufficient after a forcefull detach.