compressed bitmap: finally settle for _one_ encoding
The bitstream VLI encoding "fib 1_1" proved to be best for all collected
real-world bitmaps. It was not even second for a single real-world
bitmap. Cpu usage was not that much higher than for the whole byte based
encoding.

Drop all other encodings, as making this configurable gives no benefit,
confuses the user, and we would have a hard time explaing the differences.