GFP_KERNEL -> GFP_NOIO in various places
The asender thread must not use GFP_KERNEL, as it is in the
IO path, it is crucial for IO completion events.

The receiver thread should not use GFP_KERNEL, it usually is in the
writeout path of the peer, which would allow for distributed deadlock
if GFP_KERNEL is used.

The worker thread should usually not use GFP_KERNEL either,
as it usually is in the writeout path.

From cqueue (drbdsetup) context, we may well use GFP_KERNEL.