drbd udev rules: only install if udev version is known to work
achieved with some bash magic in scripts/Makefile
default to install the rules file as
/etc/udev/rules.d/65-drbd.rules.disabled,
and only leave off the .disabled,
if the udev version was determined to be >= 085.

if you want it to be installed (without the .disabled),
but you build from some container build environment that does not have
udevadm nor rpm nor dpkg-query (or udev not installed), you can
  make install DRBD_ENABLE_UDEV=any-non-empty-value
to force it.