fix regression: initial sync target hung in WFBitMapT
Regression was introduced with below commit, and is fixed by also
allowing WFBitMapT as expected peer connection state in receive_state.

For anything larger than WFBitMapT, we should really not consider to do
a new sync handshake, at least the peer thinks it is still busy with the
previous sync.

commit 0c739fbf298c3d274abf9d0b43225b296c2e766c
Author: Lars Ellenberg <lars.ellenberg@linbit.com>
Date:   Thu Jun 4 12:12:13 2009 +0200

    fix race in state handling which caused a full sync