drbd.ocf: improve fencing and behaviour after cluster crash
Allow fencing constraints to be placed on arbitrary node attributes,
or to be disabled.

Also unfence from monitor.
This should possibly only be done on "higher order" monitor,
to avoid to call cibadmin -Ql on every monitor.

Set minimal master-score on "Consistent" devices,
to be able to be promoted after simultaneous cluster crash.