configure.ac: add --enable-spec flag
Our use of configure aims to kill two birds with one stone: we want to
be able to create Makefiles from templates (as configure usually
does), but we also want to create distribution-specific RPM spec
files.

We need to treat these two cases separately: when we create a
spec, we don't want Makefiles, as the spec itself invokes configure
and creates the Makefiles then. Conversely, when we want Makefiles,
then we must not update the spec, as a spec shouldn't be updating
itself.

Add the --enable-spec configure flag, disabled by default. When
enabled, it creates drbd.spec from drbd.spec.in. When disabled, it
creates Makefiles from their Makefile.in templates.