drbd: better fix for spin_lock_irqsave in endio callback
This reverts the (correct, but unnecessary)
 commit 7f9c6c210158d212cc2c7be6d6b4d289078ab735
 Author: Lars Ellenberg <lars.ellenberg@linbit.com>
 Date:   Wed Nov 10 10:33:21 2010 +0100

     drbd: use irqsave in bio endio callback

In commit 9b7f76dc37919ea36caa9680a3f765e5b19b25fb,
 Author: Lars Ellenberg <lars.ellenberg@linbit.com>
 Date:   Wed Aug 11 23:40:24 2010 +0200

    drbd: new configuration parameter c-min-rate

a bad chunk slipped through, which is now reverted as well,
restoring the correct irqsave for the endio callback.

This patch also add comments at both req_mod()
and in the endio callback so it should not happen again.