build: drop drbd-rgmanager from Requires: of drbd.rpm
The whole point of the drbd.rpm virtual package is convenience, i.e.
have it require the "typical" set of subpackages of a given drbd
version. If we depend on drbd-rgmanager there, we annoy most of our
userbase by indirectly pulling in the full rgmanager stack, as the
drbd-rgmanager subpackage depends (as required by packaging guidelines)
on the rgmanager package itself.

The vast majority of our userbase uses drbd with pacemaker,
and has no use for rgmanager.