drbd.ocf: disable self-outdate by default; refer to the user's guide
Self-outdating inside the resource agent on post notifications or on
stop depends on correct OCF_RESKEY_CRM_meta_notify_master_uname,
which unfortunately has proven to be unreliable.

Until that is fixed in pacemaker, default to disabled,
and document this.

Also refer to the User's Guide for resource level fencing, seems like
people expect that automagically from this resource agent.