userland support for genetlink based configuration interface
I don't use libnl.  Using the api of libnl for the things we want to do
seemed to result in as much code (not counting the libnl code itself)
as just "opencoding" it by using a stripped down copy of various related
files from the linux kernel tree.  In total this results in less code,
and no dependency on some external library.