drbd: fix setsockopt for user mode linux
TODO: use kernel_setsockopt, and only provide a compat backport of it
for kernels that need it (< 2.6.18).

We use our own copy of kernel_setsockopt, and did not mess around with
get_fs/set_fs, since we thought we knew we would always be KERNEL_DS
anyways. Apparently not so for at least user mode linux, so put the
set_fs(KERNEL_DS) in there.