drbd: do not currupt task_struct
Don't use sprintf(current->comm, ...). We got away with this while the
only variable part in the thread name was the minor number (0-255),
and we knew it would fit in the char comm[16].
Using our arbitrary connection names now in thread names, that will
quickly result in corrupting the task struct.