userland: parser fixes for stacked resources
Introduce check_volume_sets_equal(), which replaces
ensure_vols_1_in_2(), and checks that both sets of volumes
consist of the same volume numbers.

With the special case that, if we compare a stacked resource
with it's lower resource, we allow the lower resource to
define a more volumes.

This also populates all ->disk members of all volumes of a stacked
resource (which has been a /* FIXME */ in the code).