drbd: Turn role into a resource attribute
In drbd 8.4, each volume in a resource has its own role attribute, and the role
of different volumes in a resource may differ.  The role of a resource is
determined by the volume's roles: the resource is primary if any of its volumes
is primary.  Before drbd 8.4, a volume needed to become secondary before it
could be removed.  With a per-volume role, a volume needs to become secondary
before it can be removed, even when the resource remains primary.

Per-volume roles also tie into the suspend logic: a volume suspends if it is
primary and both the local and remote disk goes out of date.  Volumes are added
to a primary resource with their role set to secondary.  They will not have any
up-to-date disks available initially, but the resource will not suspend until
the volume is made primary.

This design can be explained by drbd's history, but actually it makes little
sense.  Instead, the role should become a property of the resource.
Conceptually, there is no reason why it shouldn't be possible to remove a
"primary" volume as long as it is not in use.  A resource should
suspend only if I/O is attempted on a volume without up-to-date disks.

As a first step to fixing this, move the role to struct drbd_resource.  Suspend
and removal of volumes from primary resources will be fixed later.