drbd: drbd_adm_attach(): Wait for previously queued resource work
As described in the commit ading drbd_flush_workqueue() in drbd_adm_attach(),
the objective here is to wait for pending after-state-change work.  This work
has been moved to the resource work queue.

(The connection sender thread may not even be running and, in the future, there
may not even be a connection, yet.)