drbd: Split drbd_read_state() into drbd_get_device_state() and drbd_get_peer_device_state()
A drbd_state can represent the states of different objects (and of multiple
objects at a time) depending on the context; an additional mask clarifies what
a particular drbd_state means where the context doesn't make that clear.

In the code, we either look at the state of a device or peer device.  In the
device case, the parts of the drbd_state which refer to a peer device are
undefined.  Use different accesssor functions for the two cases.