drbd: Turn is_valid_state() into is_allowed_soft_transition()
We want to transform is_valid_state() to operate on an entire resource.
However, the way it is used in is_allowed_soft_transition() relies on the fact
that it operates on a single device: otherwise, devices in "valid" states could
become invalid because of other "invalid" devices.  So as a first step,
integrate the old state -> new state checking into is_valid_state().