drbd: Add a second set of "state" fields (1)
While state[NOW] will represent the current state, state[NEW] will represent
the desired next state during a state change.  (To improve the readability of
code that changes the state, state[OLD] can be used as an alias of state[NOW].)

This will allow to build up the new state incrementally and to finally "commit"
or "abort" the entire state change.