drbd: Get rid of new_state_tmp
The metadata of a drbd device include the MDF_CONSISTENT, MDF_WAS_UP_TO_DATE,
and MDF_PEER_OUTDATED flags which describe the last known state of the device.
When a backing device is attached (drbdsetup attach) and the peer device is
unreachable, these flags are used to determine the disk state of the drbd
device.

The new_state_tmp variable is used to hold the future disk and peer disk state
until it is clear that the peer device is unreachable.  Make this more obvious
by using two separate disk state enums instead.

If the peer device is unreachable, the MDF_PEER_OUTDATED flag also causes the
peer disk state to be set to D_OUTDATED.  When fencing is used, a node may only
become primary once it knows that its peer(s) have become outdated.  By
remembering if a peer disk has already been outdated, a fencing node may become
primary even when it cannot reach its peer(s).