drbd: Move fencing_policy from disk_conf to net_conf
We want to be able to configure fencing on a per-connection basis so that
tightly coupled nodes can fence each other without requiring fencing among all
nodes.  (Fencing among all nodes would block when nodes that haven't been
outdated become unreachable).

However, some of the connections may not yet be defined when attaching a disk,
and so the fencing policy may not be known at that point.  This means that we
cannot decide at attach time if a consistent disk can become up to date, or if
fencing prevents that.  (Only an up-to-date disk can be written to.)

To get around this problem, we add a MDF_FENCING_IND meta-data flag.  When
fencing is enabled/disabled for a connection, this flag is set/cleared for all
volumes on that connection.  We use this flag to determine if a consistent
device can become up to date when a node is started.

The on-disk meta-data format still needs to be extended to multiple
connections.  Then, MDF_FENCING_IND will become a per-connection flag.