drbd: Add peer_addr_hashes meta-data table
Within a resource, each device normally has a peer device on each peer node.
We keep track of the differences between the devices on different nodes in the
metadata: each peer device is assigned its own set of metadata based on UUIDs.

When a node is started, we may find that a device is in consistent state; this
means that the device can be promoted to primary.  However, if fencing is used,
we must first ensure that all other peer devices have been outdated.  If a peer
device is not oudated yet, we must outdate it before we can proceed.  This
requires that we know which peer devices belong to which peers, even for peer
devices which do not respond.

We maintain this mapping by remembering a hash of the network address of the
peer for each peer device.  That way, we know which peer devices belong to
which peers as soon as all connections are configured.

This will obviously fail if peer network addresses change.  In that case, it
may be necessary to outdate peer devices by other means (for example, by hand).