drbdmeta: fix upgrade path from empty 8.3 style activity log
If the 8.3 style activity log is empty, which means that node was never
been written to, it was Secondary its entire lifetime, or at least since
the last 8.3 style create-md, drbdmeta skipped re-initializing the
activity log in 8.4 style.
If that device now would be attached and detached to 8.4,
but still stay Secondary, the next apply-al and attach would fail.

Fix both cases: always re-initialize when upgrading from 8.3 to 8.4.
And only error out on "unusable activity log", if we think we actually
would need to apply it (detect a crashed former Primary).