drbd: Rewrite the bitmap code for multiple peers
For tracking the differences between a device and all its peers, we allocate
one bit for each 4k block of data per peer.  A common operation is to set or
clear all bits of a range of 4k blocks; this requires fewer I/O operations if
the bits are close together in the bitmap.  To achieve this without extreme bit
fiddling, we organize the bitmap in chunks of 32 bits: for example, with three
peers, the bits are assigned as follows:

   peer 1: bits 0-31,  96-127,  192-223, ...
   peer 2: bits 32-63, 128-159, 224-255, ...
   peer 3: bits 64-95, 160-191, 256-287, ...

The new code allocates new pages with __GFP_ZERO and sets any bits that need to
be set using the BM_OP_SET operation.