drbd: Convert drbd_bm_write_page() into drbd_bm_write_range()
The bitmap now interleaves the bits of peers, and so the bits for each peer are
no longer contiguous. This makes it non-trivial to map a range of bits of a
particular peer to a range of pages on which those bits reside.

Therefore, pass the range of bits to the bitmap code directly and let the
bitmap code figure out which pages are affected.