drbd: Add an identifier to allow to determine which notifications belong together
A single state change can affect a number of objects.  Each change of an object
results in a separate netlink notification.  Use the same identifier (nh_id in
struct drbd_notification_header) in all notifications for a particular state
change so that user space can determine which notifications belong together.