drbd: Add connection_is_alive() check
We currently create a single connection object when a resource is created.
That object becomes "alive" with "drbdsetup connect" (which sets the connection
endpoint addresses), and "dies" with "drbdsetup disconnect" (which clears the
connection endpoint addresses again).

In the future, connection objects will be created by "drbdsetup connect" and
destroyed by "drbdsetup disconnect".  As a step towards that state, check if a
connection is alive by checking if its endpoint addresses are defined; do this
in a separate function.