drbd: Improve logging
Introduce a conecept of "implicit options". There are set by drbdadm
implicitly. So far the only implicit option is the "_name" net options.
It is set to the connection name if available, otherwise to the peer's
node name.

Prepend all log lines that relate to a connection with that name.