drbd: Allocate bitmap slots already in drbd_attach_peer_device()
With patch "drbd: Add peer_addr_hashes meta-data table" the idea
was to allocate the bitmap slot later, when the connection gets
established.

But in various places we assume that a bitmap slot is known
as soon as the disk gets attached.

The other reason to allocate it early is that in case there
are no vacant slots, it is better to fail early.