drbd: auto-promote (3) New config flag auto-promote
When the new "auto-promote" flag is set, a device will automatically
promote itself when it gets opened for read-write.

When no peer is primary DRBD allows read access. It has to decline
any peer to become R_PRIMARY if we have a local process that has
a read-only open on one of the resources devices.

This avoids split brain when the drbd volume gets opened temporarily by udev
while it scans for PV signatures.

Notes on kobject_uevent():

Previously we called kobject_uevent() after every role change.
(1) Only call it when we became R_PRIMARY, since it buys
    udev nothing if it notices that it can not open the device after
    it was demoted to R_SECONDARY.
(2) Do not call it at all after role changes if auto-promote
    is enabled. Since now DRBD changes its role if udev wants to
    open us. Leaving that enabled would cause an endless loop.
Actually we should not call this after attach (as we do now) but
after we got access to any D_UP_TO_DATE data no matter if it is
after an attach or after an connect event. (But this is for an
other patch)