drbd: peer_ack (4) Receive PEER_ACK packets
With a PEER_ACK packet, the primary tells a secondary which nodes have
acknowledged a range of write requests.

To save bandwidth, the semantics of the PEER_ACK packet is, that
all the blocks previously to and the referred block were acked
by the nodes marked with acked_mask.

The PEER_ACK packet travels on the meta-data socket.