drbd: fix tl_release, it sometimes did not mark requests as barrier-acked
When we receive a P_BARRIER_ACK, we mark all corresponding requests in
that epoch as RQ_NET_DONE.

We have a paranoia check that the ack'ed barrier number is the one
we expect, which finds the oldest not yet barrier-acked request
on the transfer log.

The loop marking the requests in that epoch as done started its list
walk at the transfer log head, in error, and potentially quits
before even reaching the current epoch, if requests from an older epoch
are still on the transfer log, e.g. because of slower local IO.

Impact: sometimes, requests have not been marked as ack'ed,
and on the next P_BARRIER_ACK, the paranoia code triggers, complains
about BAD! BarrierAck #663 received, expected #665!
throws a C_PROTOCOL_ERROR, and we do a disconnec/reconnect.

Fix this by fist skipping over older epochs,
then marking all requests in the acknowledged epoch.